# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AudiModal.ai API",
    "description": "Comprehensive AI-powered document processing and analytics platform with ML/AI classification, anomaly detection, unified sync framework, and enterprise integrations.",
    "version": "1.0.0",
    "contact": {
      "name": "AudiModal.ai",
      "url": "https://audimodal.ai",
      "email": "support@audimodal.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.audimodal.ai/v1",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.audimodal.ai/v1",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:8080/v1",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/datasources": {
      "get": {
        "tags": ["Data Sources"],
        "summary": "List data sources",
        "description": "Retrieve a list of data sources for the authenticated tenant",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Filter by data source type",
            "schema": {
              "type": "string",
              "enum": ["file_upload", "sharepoint", "google_drive", "box", "dropbox", "onedrive", "confluence", "slack", "notion"]
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by sync status",
            "schema": {
              "type": "string",
              "enum": ["active", "inactive", "error", "syncing"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "datasources": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DataSource"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Data Sources"],
        "summary": "Create data source",
        "description": "Create a new data source connection",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDataSourceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Data source created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSource"
                }
              }
            }
          }
        }
      }
    },
    "/datasources/{datasourceId}": {
      "get": {
        "tags": ["Data Sources"],
        "summary": "Get data source",
        "description": "Retrieve a specific data source by ID",
        "parameters": [
          {
            "name": "datasourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSource"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Data Sources"],
        "summary": "Update data source",
        "description": "Update an existing data source",
        "parameters": [
          {
            "name": "datasourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDataSourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data source updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSource"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Data Sources"],
        "summary": "Delete data source",
        "description": "Delete a data source and stop all syncing",
        "parameters": [
          {
            "name": "datasourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Data source deleted successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "DataSource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["file_upload", "sharepoint", "google_drive", "box", "dropbox", "onedrive", "confluence", "slack", "notion"]
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "error", "syncing"]
          },
          "config": {
            "type": "object",
            "description": "Configuration specific to the data source type"
          },
          "sync_schedule": {
            "type": "object",
            "properties": {
              "enabled": {
                "type": "boolean"
              },
              "frequency": {
                "type": "string",
                "enum": ["hourly", "daily", "weekly", "monthly"]
              },
              "next_sync": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "last_sync": {
            "type": "object",
            "properties": {
              "started_at": {
                "type": "string",
                "format": "date-time"
              },
              "completed_at": {
                "type": "string",
                "format": "date-time"
              },
              "status": {
                "type": "string",
                "enum": ["success", "failed", "partial"]
              },
              "files_processed": {
                "type": "integer"
              },
              "errors": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateDataSourceRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["sharepoint", "google_drive", "box", "dropbox", "onedrive", "confluence", "slack", "notion"]
          },
          "config": {
            "type": "object",
            "description": "Configuration specific to the data source type"
          },
          "sync_schedule": {
            "type": "object",
            "properties": {
              "enabled": {
                "type": "boolean",
                "default": true
              },
              "frequency": {
                "type": "string",
                "enum": ["hourly", "daily", "weekly", "monthly"],
                "default": "daily"
              }
            }
          }
        },
        "required": ["name", "type", "config"]
      },
      "UpdateDataSourceRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "config": {
            "type": "object"
          },
          "sync_schedule": {
            "type": "object",
            "properties": {
              "enabled": {
                "type": "boolean"
              },
              "frequency": {
                "type": "string",
                "enum": ["hourly", "daily", "weekly", "monthly"]
              }
            }
          }
        }
      },
...
    }
  },
...
}